move to layout2
