+++ src/cmd/dev.go
@@ -27,7 +27,6 @@ import (
 	"github.com/zarf-dev/zarf/src/internal/packager2"
 	layout2 "github.com/zarf-dev/zarf/src/internal/packager2/layout"
 	"github.com/zarf-dev/zarf/src/pkg/archive"
-	"github.com/zarf-dev/zarf/src/pkg/layout"
 	"github.com/zarf-dev/zarf/src/pkg/lint"
 	"github.com/zarf-dev/zarf/src/pkg/logger"
 	"github.com/zarf-dev/zarf/src/pkg/message"
@@ -357,10 +356,10 @@ func newDevGenerateCommand() *cobra.Command {
 func (o *devGenerateOptions) run(cmd *cobra.Command, args []string) (err error) {
 	l := logger.From(cmd.Context())
 	name := args[0]
-	generatedZarfYAMLPath := filepath.Join(o.output, layout.ZarfYAML)
+	generatedZarfYAMLPath := filepath.Join(o.output, layout2.ZarfYAML)
 
 	if !helpers.InvalidPath(generatedZarfYAMLPath) {
-		prefixed := filepath.Join(o.output, fmt.Sprintf("%s-%s", name, layout.ZarfYAML))
+		prefixed := filepath.Join(o.output, fmt.Sprintf("%s-%s", name, layout2.ZarfYAML))
 		l.Warn("using a prefixed name since zarf.yaml already exists in the output directory",
 			"output-directory", o.output,
 			"name", prefixed)
