+++ src/internal/packager2/layout/import.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/zarf-dev/zarf/src/api/v1alpha1"
 	"github.com/zarf-dev/zarf/src/config"
-	"github.com/zarf-dev/zarf/src/pkg/layout"
 	"github.com/zarf-dev/zarf/src/pkg/zoci"
 )
 
@@ -80,7 +79,7 @@ func resolveImports(ctx context.Context, pkg v1alpha1.ZarfPackage, packagePath, 
 					return v1alpha1.ZarfPackage{}, fmt.Errorf("package %s imported in cycle by %s in component %s", filepath.ToSlash(importPath), filepath.ToSlash(packagePath), component.Name)
 				}
 			}
-			b, err := os.ReadFile(filepath.Join(importPath, layout.ZarfYAML))
+			b, err := os.ReadFile(filepath.Join(importPath, ZarfYAML))
 			if err != nil {
 				return v1alpha1.ZarfPackage{}, err
 			}
@@ -238,7 +237,7 @@ func fetchOCISkeleton(ctx context.Context, component v1alpha1.ZarfComponent, pac
 	if err != nil {
 		return "", err
 	}
-	componentDesc := manifest.Locate(filepath.Join(layout.ComponentsDir, fmt.Sprintf("%s.tar", name)))
+	componentDesc := manifest.Locate(filepath.Join(ComponentsDir, fmt.Sprintf("%s.tar", name)))
 	var tarball, dir string
 	// If the descriptor for the component tarball was not found then all resources in the component are remote
 	// In this case, we represent the component with an empty directory
